# APIDocs/Writerside/openapi.yaml
openapi: 3.0.2
servers:
  - url: /api
info:
  description: |-
    Esta API proporciona un conjunto de endpoints para gestionar trabajadores y trabajos en una empresa.
    Permite realizar operaciones CRUD en las tablas de 'trabajador' y 'trabajo', así como otras funcionalidades
    para asignar, finalizar y consultar trabajos. Además, ofrece métodos para obtener información detallada sobre
    los trabajos realizados por los trabajadores, y facilita la administración de tareas pendientes y finalizadas.
  version: 1.0.0
  title: Gestión Hotelera Spring
  contact:
    email: liam.alejoasensi@gmail.com, karlos.riverog@gmail.com
  license:
    name: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication
    url: 'https://creativecommons.org/publicdomain/zero/1.0/'
...
paths:
...
  /trabajos:
    # Servicio GET para conseguir todos los trabajos
    get:
      tags:
        - trabajo
      summary: Obtener listado de los trabajos
      description: Obtener el listado completo de todos los trabajos
      operationId: getAllTrabajos
      responses:
        '200':
          description: Devolución del listado con los trabajos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trabajo'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la consulta en la base de datos
        default:
          description: Error desconocido

    # Servicio POST para guardar un nuevo trabajo
    post:
      tags:
        - trabajo
      summary: Crear un trabajo
      description: Insertar un nuevo trabajo con datos válidos
      operationId: putTrabajo
      responses:
        '201':
          description: Trabajo creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    $ref: '#/components/schemas/Trabajo'
        '400':
          description: Intento de creación con datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: El campo 'id' no puede estar vacío
        '409':
          description: Conflicto con ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajo con el ID: 2 ya existe'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la inserción en la base de datos
        default:
          description: Error desconocido
        requestBody:
          description: El trabajo a insertar
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trabajo'  
  /trabajos/{id}:
    # Servicio GET para conseguir los datos de un trabajo por ID
    get:
      tags:
        - trabajo
      summary: Obtener trabajo por ID
      description: Devuelve el detalle de un trabajo según su ID
      operationId: getTrabajoById
      responses:
        '200':
          description: Datos del trabajo devueltos exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    $ref: '#/components/schemas/Trabajo'
        '404':
          description: El trabajo no fue encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajo con el id: T3 no se ha encontrado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la consulta en la base de datos
        default:
          description: Error desconocido
    post:
      tags:
        - trabajo
      summary: Crear un trabajo con un trabajador
      description: |-
        Se intenta buscar el trabajador según el ID pasado por parámetros y se crea un trabajo con el trabajador
        asignado a dicho trabajo directamente.
      operationId: postTrabajoAsignado
      responses:
        '201':
          description: Trabajo creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    $ref: '#/components/schemas/Trabajo'
        '400':
          description: Intento de creación con datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: El campo 'id' no puede estar vacío
        '404':
          description: El trabajador no fue encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajador con el id: 3 no se ha encontrado'
        '409':
          description: Conflicto con ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajo con el ID: T2 ya existe'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la inserción en la base de datos
        default:
          description: Error desconocido
    # Servicio PUT para actualizar los datos de un trabajador
    put:
      summary: Actualiza un trabajo
      description: Busca un trabajo por ID y actualiza sus datos
      operationId: putTrabajo
      parameters:
        - in: path
          name: id
          required: true
          description: ID del trabajo
          schema:
            type: string
      requestBody:
        description: El trabajo a actualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trabajo'
      responses:
        '201':
          description: Trabajo actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    $ref: '#/components/schemas/Trabajo'
        '400':
          description: Intento de actualización con datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: El campo 'id' no puede estar vacío
        '404':
          description: El trabajo no fue encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajo con el id: T3 no se ha encontrado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al actualizar el trabajo en la base de datos
        default:
          description: Error desconocido
    # Servicio DELETE para eliminar un trabajador según su ID
    delete:
      summary: Elimina un trabajo
      description: Busca un trabajo por ID y lo elimina de la BBDD
      operationId: deleteTrabajo
      parameters:
        - in: path
          name: id
          required: true
          description: ID del trabajo
          schema:
            type: string
      responses:
        '200':
          description: Trabajo eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    type: string
                    example: Trabajo eliminado con éxito
        '404':
          description: El trabajo no fue encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajo con el id: T3 no se ha encontrado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al eliminar el trabajo de la base de datos
        default:
          description: Error desconocido
  /trabajos/sin_asignar:
    get:
      summary: Obtiene los trabajos sin asignar
      description: Busca todos los trabajos y devuelve aquellos que no tengan un trabajador asignado
      operationId: getTrabajosSinAsignar
      responses:
        '200':
          description: Devolución del listado con los trabajos sin asignar
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trabajo'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la consulta en la base de datos
        default:
          description: Error desconocido
  /trabajos/{trabajadorId}/finalizados/{fecIni}/{fecFin}:
    get:
      summary: Obtiene los trabajos finalizados entre dos fechas
      description: |-
        Busca todos los trabajos realizados por un trabajador (Que se busca por el ID pasado por parámetros) y que se
        hayan realizado entre la fecha de inicio (Primera fecha en los parámetros) y la fecha de fin (Segunda fecha en
        los parámetros), ambas inclusive.
      operationId: getTrabajosEntreFechas
      parameters:
        - in: path
          name: trabajadorId
          required: true
          description: ID del trabajador cuyas tareas se quieren buscar
          schema:
            type: string
        - in: path
          name: fecIni
          required: true
          description: Fecha de inicio de la búsqueda
          schema:
            type: string
            format: date
        - in: path
          name: fecFin
          required: true
          description: Fecha de fin de la búsqueda
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Devolución del listado con los trabajos finalizados entre las dos fechas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trabajo'
        '400':
          summary: Error relacionado con las fechas
          description: |-
            Mensaje de error en caso de que las fechas no tengan el formato esperado (d-MM-yyyy), o de que se indique
            la fecha de inicio como una fecha posterior a la fecha de fin
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: La fecha de inicio no puede ser posterior a la fecha de fin
        '404':
          description: El trabajador no fue encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: 'El trabajador con el id: 3 no se ha encontrado'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la consulta en la base de datos
        default:
          description: Error desconocido
  /trabajos/finalizados:
    # Servicio GET para conseguir todos los trabajos
    get:
      tags:
        - trabajos_finalizados
      summary: Obtener listado de los trabajos realizados
      description: Obtener el listado completo de todos los trabajos finalizados por todos los trabajadores.
      operationId: getAllTrabajosFinalizados
      responses:
        '200':
          description: Devolución del listado con los trabajos finalizados
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trabajo'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  errorMessage:
                    type: string
                    example: Error al realizar la consulta en la base de datos
        default:
          description: Error desconocido
...
components:
  schemas:
...
    Trabajo:
      properties:
        codTrab:
          type: string
          example: T1
        categoria:
          $ref: '#/components/schemas/Especialidad'
        descripcion:
          type: string
          example: Limpieza de la habitación 102
        fecIni:
          type: string
          format: date
        fecFin:
          type: string
          format: date
          nullable: true
        tiempo:
          type: number
          format: float
          description: Horas empleadas en el trabajo
          nullable: true
          example: 1.5
        prioridad:
          type: int
          format: int32
          minimum: 1
          maximum: 4
          description: Importancia del trabajo (Del 1 al 4)
          example: 2
        idTrabajador:
          type: string
          description: ID del trabajador asociado al trabajo
          example: 2
      type: object
    Especialidad:
      type: string
      enum:
        - electricidad
        - limpieza
        - gestion
        - fontaneria
        - carpinteria
      example: limpieza
...